Move chat and complete to `vllm.cmd.openai`
