+++ vllm/cmd/types.py
@@ -15,7 +15,7 @@ class CLISubcommand:
         raise NotImplementedError("Subclasses should implement this method")
 
     def validate(self, args: argparse.Namespace) -> None:
-        # No validation by deafult
+        # No validation by default
         pass
 
     def subparser_init(
